Step 2: Install the SlackRubyBot gem
We will be using the gem slack-ruby-bot to build our bot. It deals
with all the nitty gritty details of Slack messaging for us, so we
can focus on the fun stuff.

In addition, we need an asynchronous messaging bus to communicate
with our Slack workspace in real-time. We're using `async-websocket`
for this.